Modify file structure, add README
