+++ README.md
@@ -1,0 +1,52 @@
+# Spot Helper
+
+## Description
+
+## Prerequisites
+
+1. Ensure that the following are satisfied before running:
+* Python (3.11 or greater) is installed on your machine.
+* GUI displays are enabled on your machine (i.e., X11 forwarding is enabled for remote connections)
+
+1. (Optional) Create a python virtual environment:
+
+	```
+	python3 -m venv venv
+	source venv/bin/activate
+	```
+
+
+1. Install python dependencies:
+
+	```
+	pip install -r requirements.txt
+	```
+ 
+
+## Usage
+
+### Linux
+
+1. Run the application:
+
+	```
+	python src/spot_helper.py
+	```
+
+### Windows
+
+1. Package the project using PyInstaller:
+
+	```
+	# TODO: Jimmy to fill this part in
+	```
+
+1. ???
+
+1. Run the application:
+
+	```
+	# TODO: Jimmy to fill this part in
+	```
+
+1. Profit?
