+++ requirements.in
@@ -2,3 +2,4 @@ matplotlib==3.10.3
 numpy==2.3.0
 pandas==2.3.0
 requests==2.32.4
+scipy==1.15.3
