+++ requirements.txt
@@ -14,6 +14,7 @@ pyparsing==3.2.3
 python-dateutil==2.9.0.post0
 pytz==2025.2
 requests==2.32.4
+scipy==1.15.3
 six==1.17.0
 tzdata==2025.2
 urllib3==2.5.0
